requirements.txt
label-studio==1.8.0
transformers==4.30.2
torch==2.0.1
Pillow==10.0.0
numpy==1.24.3 